Ready to post to render

+++ controllers/smartCanController.js
@@ -41,6 +41,7 @@ const deleteACan = async (req, res, next) => {
     // delete a smart can
     try {
         const { id } = req.query;
+        if(!id) throw new Error('Please provide a id');
         await SmartCan.deleteOne({_id: id});
         res.status(204).json({mes: `Smart can with id ${id} has been deleted`, success: true});
     } catch (error) {
@@ -55,6 +56,7 @@ const getFullState = async (req, res, next) => {
     // Getting full state
     try {
         const { id } = req.query;
+        if(!id) throw new Error('Please provide a id');
         const myCan = await SmartCan.findById(id);
         if(!myCan) throw new Error('Couldn\'t find the requested can');
         res.json({is_full: myCan.is_full, success: true});
@@ -67,6 +69,7 @@ const switchingFullState = async (req, res, next) => {
     // Switching the full state
     try {
         const { id } = req.query;
+        if(!id) throw new Error('Please provide an id');
         const myCan = await SmartCan.findById(id);
         if(!myCan) throw new Error('Couldn\'t find the requested can');
         if(myCan.is_full) myCan.is_full = false;
@@ -78,4 +81,18 @@ const switchingFullState = async (req, res, next) => {
     }
 };
 
-module.exports = {getFullState, switchingFullState, getAllCans, createCans, deleteACan}
+const updateSmartCanName = async (req, res, next) => {
+    try {
+        const {id, newName} = req.query;
+        if(!id || !newName) throw new Error('Please provide both an id and a name');
+        const foundCan = await SmartCan.findById(id);
+        if(!foundCan) throw new Error('Couldn\'t find the requested can with id ' + id);
+        foundCan.name = newName;
+        await foundCan.save();
+        res.status(201).json({mes: `The smart can with id of ${id}'s name updated to ${newName}`});
+    } catch (error) {
+        next(error);
+    }
+}
+
+module.exports = {getFullState, switchingFullState, getAllCans, createCans, deleteACan, updateSmartCanName}
